refactor: entity directory structure 재구성
