+++ src/main/java/com/rdc/weflow_server/entity/company/CompanyStatus.java
@@ -1,4 +1,4 @@
-package com.rdc.weflow_server.entity.user;
+package com.rdc.weflow_server.entity.company;
 
 public enum CompanyStatus {
     ACTIVE, INACTIVE;
