+++ src/main/java/com/rdc/weflow_server/entity/project/ProjectMember.java
@@ -1,6 +1,7 @@
-package com.rdc.weflow_server.entity.user;
+package com.rdc.weflow_server.entity.project;
 
 import com.rdc.weflow_server.entity.BaseEntity;
+import com.rdc.weflow_server.entity.user.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
