+++ src/main/java/com/rdc/weflow_server/entity/project/ProjectStatus.java
@@ -1,4 +1,4 @@
-package com.rdc.weflow_server.entity.user;
+package com.rdc.weflow_server.entity.project;
 
 public enum ProjectStatus {
     CONTRACT, IN_PROGRESS, DELIVERY, MAINTENANCE, CLOSED
